requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.30.3
langchain==0.2.3
langchain-openai==0.1.8
langgraph==0.1.1
chromadb==0.5.0
pytest==8.2.1
httpx==0.27.0
pytest-asyncio==0.23.7
